refactor(numerical-semigroups): name split contract modules for CI planning
