+++ src/jacobian/math/numerical_semigroups/_factorization_operations.py
