+++ src/jacobian/math/numerical_semigroups/_presentation_operations.py
