+++ src/jacobian/math/numerical_semigroups/_summary_operations.py
