+++ tests/math/numerical_semigroups/test_extended.py
@@ -4,7 +4,7 @@ import pytest
 from pydantic import ValidationError
 from tests.math.numerical_semigroups._support import numerical_semigroup_error
 
-from jacobian.math.numerical_semigroups._factorizations import (
+from jacobian.math.numerical_semigroups._factorization_operations import (
     FactorizationComputeRequest,
     FactorizationComputeResult,
     FactorizationDistanceRequest,
@@ -42,7 +42,7 @@ from jacobian.math.numerical_semigroups._operations import (
     compute_element_delta_set,
     compute_element_elasticity,
 )
-from jacobian.math.numerical_semigroups._presentations import (
+from jacobian.math.numerical_semigroups._presentation_operations import (
     MinimalPresentationRequest,
     MinimalPresentationResult,
     PresentationBinomialsRequest,
